refactor(day09): update loop indices to size_t for consistency
Modified loop indices in part1.cpp and part2.cpp to use size_t instead of int for better compatibility with container sizes. This change enhances type safety and prevents potential signed/unsigned comparison issues.

Key changes:
- Updated loop declarations in both part1.cpp and part2.cpp to use size_t for iterating over string and vector sizes.

+++ day09/part1.cpp
@@ -11,7 +11,7 @@ int main() {
   vector<int> blocks;
   int fileId = 0;
 
-  for (int i = 0; i < s.length(); i++) {
+  for (size_t i = 0; i < s.length(); i++) {
     int len = s[i] - '0';
     if (i % 2 == 0) {
       for (int j = 0; j < len; j++) {
@@ -34,7 +34,7 @@ int main() {
       break;
 
     int left = 0;
-    while (left < blocks.size() && blocks[left] != -1) {
+    while (static_cast<size_t>(left) < blocks.size() && blocks[left] != -1) {
       left++;
     }
     if (left >= right)
@@ -45,7 +45,7 @@ int main() {
   }
 
   long long checksum = 0;
-  for (int i = 0; i < blocks.size(); i++) {
+  for (size_t i = 0; i < blocks.size(); i++) {
     if (blocks[i] != -1) {
       checksum += (long long)i * blocks[i];
     }
